+++ src/events/SNS/sns.js
@@ -1,30 +1,15 @@
-import Promise from 'bluebird';
-
 import LambdaEvent from '../lambdaEvent';
 import { OK, ERROR } from '../../global';
 import Record from './record';
-import Response from './response';
 
 export default class extends LambdaEvent {
-  constructor(event, context, cb) {
-    super();
-    /**
-     * The raw event received by Lambda
-     * @member {Object} SNS#event
-     */
-    this.event = event;
-    this.records = event.Records.map(r => new Record(r));
-    /**
-     * The raw context object received by Lambda
-     * @member {Object} SNS#context
-     */
-    this.context = context;
-    /**
-     * The response object allowing any extending class
-     * to respond to SNS.
-     * @member {Response} SNS#response
-     */
-    this.response = new Response(event, cb);
+  /**
+   * The map of records received by the event.
+   *
+   * @return {Array[Record]} array of Record objects
+   */
+  get records() {
+    return this.event.Records.map(r => new Record(r));
   }
 
   each() {