+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.0
+
+- Added DynamoDB Events
+- Added SNS Events
+- Removed `this.response` and implemented `this.respond(OK | ERROR)`
+
 # 0.0.1
 
 - Initial Implementation